Add an option to select the floating-point precision

+++ embeddings.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2016  Mikel Artetxe <artetxem@gmail.com>
+# Copyright (C) 2016-2018  Mikel Artetxe <artetxem@gmail.com>
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -16,21 +16,21 @@
 import numpy as np
 
 
-def read(file, threshold=0, vocabulary=None):
+def read(file, threshold=0, vocabulary=None, dtype='float'):
     header = file.readline().split(' ')
     count = int(header[0]) if threshold <= 0 else min(threshold, int(header[0]))
     dim = int(header[1])
     words = []
-    matrix = np.empty((count, dim)) if vocabulary is None else []
+    matrix = np.empty((count, dim), dtype=dtype) if vocabulary is None else []
     for i in range(count):
         word, vec = file.readline().split(' ', 1)
         if vocabulary is None:
             words.append(word)
-            matrix[i] = np.fromstring(vec, sep=' ')
+            matrix[i] = np.fromstring(vec, sep=' ', dtype=dtype)
         elif word in vocabulary:
             words.append(word)
-            matrix.append(np.fromstring(vec, sep=' '))
-    return (words, matrix) if vocabulary is None else (words, np.array(matrix))
+            matrix.append(np.fromstring(vec, sep=' ', dtype=dtype))
+    return (words, matrix) if vocabulary is None else (words, np.array(matrix, dtype=dtype))
 
 
 def write(words, matrix, file):
